Move BlockProviderExecutor in status handling source, in preparation for merging that class with status handling executor class
